Simplify and rename test-project to hello-world
