+++ packages/hello-world/README.md
@@ -1,0 +1,22 @@
+# Hello World
+
+This is a test app for Managed Runtime, for internal use. 
+
+This app purposefully has the smallest footprint possible, to help 
+Salesforce staff talk about and test the feature-set of Managed Runtime
+itself, without being distracted by the complex dependencies of
+eg. the Retail React App.
+
+Please do not add features to this app to make it an attractive starting
+point for developer projects! We need something minimal for internal use.
+Start a brand-new project template instead, if there's need for one.
+
+This app is *intentionally not* a PWA – MRT does not require it.
+
+## Setup/Run/Deploy
+
+```bash
+npm ci
+npm start
+npm run push
+```
