+++ packages/pwa-kit-create-app/.gitignore
@@ -1,5 +1,5 @@
 generated-project
 local-npm-repo/storage
 template
-template-minimal
+template-hello-world
 node_modules
