+++ packages/pwa-kit-create-app/package.json
@@ -12,7 +12,7 @@
   "files": [
     "assets",
     "template",
-    "template-minimal"
+    "template-hello-world"
   ],
   "scripts": {
     "prepare": "node scripts/build.js",
