+++ packages/pwa-kit-create-app/scripts/build.js
@@ -14,13 +14,13 @@ sh.config.silent = false
 
 const monorepoRoot = p.resolve(__dirname, '..', '..', '..')
 const templatePath = p.resolve(__dirname, '..', 'template')
-const minimalTemplatePath = p.resolve(__dirname, '..', 'template-minimal')
+const helloWorldTemplatePath = p.resolve(__dirname, '..', 'template-hello-world')
 
 const mkdtempSync = () => fs.mkdtempSync(p.resolve(os.tmpdir(), 'pwa-template-tmp'))
 
 const main = () => {
     sh.rm('-rf', templatePath)
-    sh.rm('-rf', minimalTemplatePath)
+    sh.rm('-rf', helloWorldTemplatePath)
 
     const tmpDir = mkdtempSync()
     const checkoutDir = p.join(tmpDir, 'mobify-platform-sdks')
@@ -31,7 +31,7 @@ const main = () => {
     )
 
     sh.cp('-R', p.join(checkoutDir, 'packages', 'pwa'), templatePath)
-    sh.cp('-R', p.join(checkoutDir, 'packages', 'pwa-test-app'), minimalTemplatePath)
+    sh.cp('-R', p.join(checkoutDir, 'packages', 'hello-world'), helloWorldTemplatePath)
     sh.rm('-rf', tmpDir)
 }
