+++ packages/pwa-kit-create-app/template-minimal/.gitignore
@@ -1,0 +1,3 @@
+dist
+node_modules
+build.tar
