+++ packages/pwa-kit-create-app/template-minimal/.prettierignore
@@ -1,0 +1,4 @@
+build
+docs
+coverage
+scripts/generator/assets
