+++ packages/pwa-kit-create-app/template-minimal/.prettierrc.yaml
@@ -1,0 +1,6 @@
+printWidth: 100
+singleQuote: true
+semi: false
+bracketSpacing: false
+tabWidth: 4
+arrowParens: 'always'
