+++ packages/pwa-kit-create-app/template-minimal/README.md
