+++ packages/pwa-kit-create-app/template-minimal/loader.js
@@ -1,0 +1,2 @@
+// Intentionally empty.
+// This unused file is required to upload a bundle.
