+++ packages/pwa-kit-create-app/template-minimal/package.json
@@ -1,0 +1,37 @@
+{
+    "name": "pwa-test-app",
+    "version": "0.11.0-dev",
+    "license": "See license in LICENSE",
+    "private": true,
+    "dependencies": {
+        "aws-serverless-express": "^3.3.6",
+        "express": "^4.17.1"
+    },
+    "devDependencies": {
+        "@babel/cli": "^7.14.3",
+        "@babel/core": "^7.14.3",
+        "@babel/runtime": "^7.14.0",
+        "prettier": "1.18.2",
+        "pwa-kit-react-sdk": "^0.11.0-dev"
+    },
+    "scripts": {
+        "format": "prettier --write \"**/*.{js,jsx}\"",
+        "start": "node ssr.js",
+        "build": "rm -rf dist && mkdir dist && cp -R $(find . -depth 1 ! -iname 'node_modules' ! -iname 'dist' ! -iname '.*' ! -iname 'build.tar') dist/ && cd dist && npm install --production",
+        "push": "npm run build && sdk-upload --buildDirectory dist --projectSlug $npm_package_name",
+        "save-credentials": "sdk-save-credentials"
+    },
+    "mobify": {
+        "ssrEnabled": true,
+        "ssrParameters": {
+            "ssrFunctionNodeVersion": "14.x"
+        },
+        "ssrOnly": [
+            "ssr.js",
+            "node_modules/**/*.*"
+        ],
+        "ssrShared": [
+            "intentionally-does-not-exist"
+        ]
+    }
+}
