--- packages/pwa-test-app/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-dist
-build.tar
